+++ src/main/kotlin/com/github/inxilpro/intellijalpine/AttributesProvider.kt
@@ -1,7 +1,6 @@
 package com.github.inxilpro.intellijalpine
 
 import com.intellij.psi.PsiElement
-import com.intellij.psi.meta.PsiPresentableMetaData
 import com.intellij.psi.xml.XmlTag
 import com.intellij.util.ArrayUtil
 import com.intellij.xml.XmlAttributeDescriptor
@@ -34,12 +33,12 @@ class AttributesProvider : XmlAttributeDescriptorsProvider {
         return null
     }
 
+    // Dropping PsiPresentableMetaData for now
     private class AttributeDescriptor(private val name: String) :
-        BasicXmlAttributeDescriptor(),
-        PsiPresentableMetaData {
-        override fun getIcon() = Alpine.ICON
+        BasicXmlAttributeDescriptor() {
+        // override fun getIcon() = Alpine.ICON
 
-        override fun getTypeName(): String? = "Alpine.js"
+        // override fun getTypeName(): String? = "Alpine.js"
 
         override fun init(psiElement: PsiElement) {
         }
@@ -50,7 +49,7 @@ class AttributesProvider : XmlAttributeDescriptorsProvider {
         override fun isEnumerated(): Boolean = false
         override fun getDeclaration(): PsiElement? = null
         override fun getName(): String = name
-        override fun getDependences(): Array<Any> = ArrayUtil.EMPTY_OBJECT_ARRAY
+        override fun getDependencies(): Array<Any> = ArrayUtil.EMPTY_OBJECT_ARRAY
         override fun isFixed(): Boolean = false
         override fun getDefaultValue(): String? = null
         override fun getEnumeratedValues(): Array<String>? = ArrayUtil.EMPTY_STRING_ARRAY